# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UploadResult {
  success: boolean;
  filename: string;
  publicUrl: string;
}

export interface PhotoItem {
  name: string;
  publicUrl: string;
  /** Proxied preview URL served by our backend for displaying in img tags */
  previewProxyUrl: string;
}

export interface PhotosList {
  article: string;
  photos: PhotoItem[];
}

export interface CatalogFolder {
  article: string;
  photoCount: number;
  /** Proxied URL of the first photo for use as thumbnail */
  coverProxyUrl: string;
}

export interface CatalogResult {
  folders: CatalogFolder[];
}

export interface ErrorResult {
  error: string;
}

export type UploadPhotoBody = {
  /** Article number (folder name on Yandex Disk) */
  article: string;
  /** Photo file */
  photo: Blob;
};