# lib/api-zod/src/generated/types/catalogFolder.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CatalogFolder {
  article: string;
  photoCount: number;
  /** Proxied URL of the first photo for use as thumbnail */
  coverProxyUrl: string;
}