# lib/api-zod/src/generated/types/catalogResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CatalogFolder } from "./catalogFolder";

export interface CatalogResult {
  folders: CatalogFolder[];
}