# lib/api-zod/src/generated/types/photoItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PhotoItem {
  name: string;
  publicUrl: string;
  /** Proxied preview URL served by our backend for displaying in img tags */
  previewProxyUrl: string;
}