# lib/api-zod/src/generated/types/photosList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PhotoItem } from "./photoItem";

export interface PhotosList {
  article: string;
  photos: PhotoItem[];
}